refactoring: Adicionando tratamento de erros na camada de serviço de Tabela

+++ src/services/tabelaService.js
@@ -1,4 +1,5 @@
 const Tabela = require('./../entities/tabela')
+const Mensagens = require('../mensagens')
 
 class TabelaService {
 
@@ -11,12 +12,10 @@ class TabelaService {
   }
 
   async find(idTabela) {
-    if(idTabela) {
-      if(!isNaN(idTabela)) {
-        return await this.tabelaRepository.find(idTabela)
-      } else {
-        return {message: `o id informado é inválido!`}
-      }
+    if(!isNaN(idTabela)) {
+      return await this.tabelaRepository.find(idTabela)
+    } else {
+      throw new Error(Mensagens.PARAMETRO_INVALIDO)
     }
   }
 
@@ -26,15 +25,23 @@ class TabelaService {
       return await this.tabelaRepository.create(tabela)
     }
 
-    return {message: tabela.isValid().campos}
+    throw new Error(tabela.isValid().property.shift())
   }
 
   async update(data) {
+    if(data.id) {
       return await this.tabelaRepository.update(data)
+    }
+
+    throw new Error(Mensagens.PARAMETRO_ID_OBRIGATORIO)
   }
 
   async delete(idTabela) {
-    return await this.tabelaRepository.delete(idTabela)
+    if(!isNaN(idTabela)) {
+      return await this.tabelaRepository.delete(idTabela)
+    }
+    
+    throw new Error(Mensagens.PARAMETRO_INVALIDO)
   }
 }
